refactor(lib): split lib.ts into consts, utils, and core API modules
Move DEFAULT_TRIGGERS/DEFAULT_DESCRIBE_OPTIONS/DEFAULT_TEST_ACTOR_OPTIONS to
consts.ts and actor-run helpers (createStartRunFn, createStartStandbyFn,
createStandbyTask, generateRunLink) to utils.ts. lib.ts now contains only
module-level init, the trigger stack, and the exported API (describe,
testActor, testStandbyActor). Also inlines BACKWARD_COMPATIBLE_HOURLY_DIR
as process.env access so it is no longer exported from index.ts.
